+++ tfe/provider.go
@@ -113,6 +113,7 @@ func Provider() *schema.Provider {
 			"tfe_oauth_client":            dataSourceTFEOAuthClient(),
 			"tfe_organization_membership": dataSourceTFEOrganizationMembership(),
 			"tfe_organization_run_task":   dataSourceTFEOrganizationRunTask(),
+			"tfe_organization_tags":       dataSourceTFEOrganizationTags(),
 			"tfe_slug":                    dataSourceTFESlug(),
 			"tfe_ssh_key":                 dataSourceTFESSHKey(),
 			"tfe_team":                    dataSourceTFETeam(),